% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvoBind.R
\name{AgeBind}
\alias{AgeBind}
\title{Internal function to graft at specific time point
len = point to graft (in units of branch lengths);
mindist = branch length if len is too young/old for
the branch in question; terminal = is grafting using
a terminal branch}
\usage{
AgeBind(tree, node, sp_name, len, mindist, terminal)
}
\description{
Internal function to graft at specific time point
len = point to graft (in units of branch lengths);
mindist = branch length if len is too young/old for
the branch in question; terminal = is grafting using
a terminal branch
}
