% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvoPhylo.R
\name{AvoPhylo}
\alias{AvoPhylo}
\title{AvoPhylo: Building phylogenys' based on AvoTrex extinct bird trait database and BirdTree backbone}
\usage{
AvoPhylo(ctrees, avotrex, PER = 0.2, tax, Ntree, n.cores = 1, cluster.ips = NULL)
}
\arguments{
\item{avotrex}{The AvoTrex database of extinct species phylogeny and taxonomic relationships. This data contains the orders to append each species to the supplied extant phylogenies in a set order.}

\item{PER}{Precentage/fraction for branch truncation based on random grafting (see AvoBind for more details).}

\item{tax}{The Jetz et al. (2012) taxonomy .csv. Supplied as data within the package.}

\item{Ntree}{The number of trees to sample from the supplied number of trees. Value must be greater than the number of supplied trees (ctrees)}

\item{n.cores}{Number of cores used to build the phylogeny. Default is one (will run with parallel processing)}

\item{cluster.ips}{Cluster location. Keep as default.}

\item{ctree}{multiPhylo object containing phylogenies.}
}
\value{
The function returns a multiPhylo object consisting of N trees that were randomly selected from a supplied number. These trees have all had the extinct species from AvoTrex grafted on. For more details on the grafting, see: **PAPER**
}
\description{
AvoPhylo: Building phylogenies based on AvoTrex extinct bird trait database and BirdTree backbone
}
\details{
Package to build phylogenies incorporating the extinct species from the AvoTrex extinct birds database.
}
\examples{
# trees <- AvoPhylo(ctrees = ctrees, avotrex = avotrex, PER = 0.2, tax = taxonomy, Ntree = 1, n.cores = 1, cluster.ips = NULL)
}
