% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_BirdTree_taxa.R
\docType{data}
\name{BirdTree_tax}
\alias{BirdTree_tax}
\title{Extant species taxonomy from Jetz et al. (2012)}
\format{
An object of class \code{data.frame} with 9993 rows and 11 columns.
}
\usage{
data(BirdTree_tax)
}
\description{
Extant species taxonomy from Jetz et al. (2012)
}
\examples{
data(BirdTree_tax)
}
\references{
Jetz, W., Thomas, G. H., Joy, J. B., Hartmann, K., & Mooers, A. O. (2012). The global diversity of birds in space and time. Nature, 491(7424), 444-448.
}
\keyword{datasets}
