\name{BirdTree_trees}
\alias{BirdTree_trees}
\docType{data}
\title{
Bird Tree trees
}
\description{
Two Bird Tree backbone trees obtained from Jetz et al. (2012). 
}
\usage{data("BirdTree_trees")}
\format{
  A multiPhylo object consisting of two elements. 
}
\details{
If necessary, more details than the __description__ above
}
\source{
Jetz, W., Thomas, G., Joy, J. et al. The global diversity of birds in space and time. Nature 491, 444–448 (2012). https://doi.org/10.1038/nature11631
}
\references{
Jetz, W., Thomas, G., Joy, J. et al. The global diversity of birds in space and time. Nature 491, 444–448 (2012). https://doi.org/10.1038/nature11631
}
\examples{
data(BirdTree_trees)
}
\keyword{datasets}
