% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvoBind.R
\name{AvoBind}
\alias{AvoBind}
\title{AvoBind - Internal function}
\usage{
AvoBind(tree, node, per, per_fixed = FALSE, sp_name)
}
\arguments{
\item{tree}{Tree object (i.e., phylogeny)}

\item{node}{Target node}

\item{per}{The fraction (0-1) of total branch length to truncate at either
end of the branch for grafting (e.g. 0.2 cuts of 20% of the total branch
length from either end)}

\item{per_fixed}{whether to graft on at an exact place rather than random;
value between 0-1, with larger number meaning grafting happens closer to
the root.}

\item{sp_name}{Name of the grafted species}
}
\description{
AvoBind - Internal function
}
