% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvoPhylo.R
\name{AvoPhylo}
\alias{AvoPhylo}
\title{AvoPhylo: Building phylogenies based on AvoTrex extinct bird trait database and BirdTree backbone}
\usage{
AvoPhylo(ctrees, avotrex, PER = 0.2, tax, Ntree, n.cores = 1, cluster.ips = NULL)
}
\arguments{
\item{ctrees}{multiPhylo object containing phylogenies.}

\item{avotrex}{The AvoTrex extinct species phylogeny database. This database
contains the information and commands required to graft the extinct species
on to the BirdTree trees.}

\item{PER}{Percentage/fraction for branch truncation based on random grafting
(see AvoBind for more details).}

\item{tax}{The Jetz et al. (2012) BirdTree taxonomy .csv. Supplied as data
within the package.}

\item{Ntree}{The number of trees to sample from the supplied number of
BirdTree trees. Value must be greater than the number of supplied trees
(ctrees)}

\item{n.cores}{Number of cores used to build the phylogeny. Default is one
(will run with parallel processing)}

\item{cluster.ips}{Cluster location. Keep as default.}
}
\value{
The function returns a multiPhylo object consisting of N trees that
  were randomly selected from a supplied number. These trees have all had the
  extinct species from AvoTrex grafted on. For more details on the grafting,
  see: Sayol et al. (IN PREP) The global loss of avian functional and phylogenetic diversity 
  from extinctions in the Holocene and Late Pleistocene

The imputed tree(s) with the extinct species grafted on. The object is returned as a class 'multiPhylo', the same 
as the input tree(s).
}
\description{
AvoPhylo: Building phylogenies based on the AvoTrex database and BirdTree
backbone
LazyData: true
}
\details{
Function to build phylogenies incorporating the extinct species from the AvoTrex extinct birds database (Sayol et al.).
AvoTrex provides data on geographical location, island endemicity, volancy, body size and standard external and skeleton 
morphological measurements for 602 extinct bird species. The AvoPhylo function provides a pipeline to incorporate the 
extinct species from AvoTrex into the "BirdTree" phylogeny of exant birds (Jetz et al. 2012). Utilising codes assigned to 
each species based on their known taxonomic affinities, the function binds each species in turn to provided phylogenetic 
trees. 

The species are grafted onto the tree in a set order provided in the column " phylo_id", as certain species need to
be grafted onto the tree before other species. Some species are assigned to groups within the data. These species are 
assigned a code "xS" within the column "phylo_id2". These species groups consist of close relatives, whose exact taxonomic 
relationships are unknown. Therefore, the order in which they are joined is randomised. 

As some of the codes (see table below) randomly place the given species within a group of species, a genus, or a family, 
and some species groups are randomised before grafting (see above), it is useful to run the grafting proceedure over a 
a number of trees to average out the randomisation. Therefore, the function can be run in parallel using the argument
"n.cores". Note that the function will run on one core as default and if only one tree is supplied. Trees can also be 
randomly selected from a number of trees by giving the function a group of trees using the argument "ctrees" and then 
defining a smaller number using "Ntree". If the maximum number of trees is to be used, "ctrees" and "Ntree" should have 
the same value. 

Codes | Full name                   | Definition                                                                          |
--------------------------------------------------------------------------------------------------------------------------|
S     | Sister                      | Grafted as a sister to a known extant or extinct species already in the tree        |
--------------------------------------------------------------------------------------------------------------------------|
SSG   | Sister species group        | Grafted as a sister to a group of extant and/or extinct species already in the tree |
--------------------------------------------------------------------------------------------------------------------------|
SGG   | Sister genus group          | Grafted as a sister to an entire extant or extinct genus (i.e., for the first       |
      |                             | grafted representative of an extinct genus)                                         |
--------------------------------------------------------------------------------------------------------------------------|      
SGG2  | Sister genus group 2        | Grafted as sister to multiple genera. This was for when a species was sister to a   |
      |                             | subfamily or some other large specific clade                                        |
--------------------------------------------------------------------------------------------------------------------------|    
SFG   | Sister family group         | Grafted as a sister to an entire extant or extinct family already present in the    |
      |                             | tree (i.e., for the first grafted representative of an extinct family)              |
--------------------------------------------------------------------------------------------------------------------------|     
SOG   | Sister order group          | Grafted as a sister to an entire order already present in the tree (i.e., for the   |
      |                             | first grafted representative of an extinct order)                                   |
--------------------------------------------------------------------------------------------------------------------------|     
RSG   | Random species group        | Grafted to a randomly selected species from a pre-defined group of species (i.e.,   | 
      |                             | from which is believed to have close affinities                                     |
--------------------------------------------------------------------------------------------------------------------------|     
RGG   | Random genus group          | Grafted to a randomly selected species from a given genus. For example, if an       |
      |                             | extinct species was believed to be a finch derived from a European finch species,   |
      |                             | but the exact sister species is unknown.                                            | 
--------------------------------------------------------------------------------------------------------------------------|                                     
RGG2  | Random genus group 2        | Grafted to a randomly selected species from a group of genera (e.g. when all that is| 
      |                             | known is that the species is from a specific subfamily). Currently not used in the  |
      |                             | database, but the relevant functionality has been kept in the R script, as it could |
      |                             | be useful for future studies                                                        | 
--------------------------------------------------------------------------------------------------------------------------|                       
RFG   | Random family group         | Grafted to a randomly selected species from a given family                          |  
--------------------------------------------------------------------------------------------------------------------------|                                                                                                
RSGG  | Random sister genus group   | Grafted as sister to a randomly selected genus from a pre-defined group of genera   | 
--------------------------------------------------------------------------------------------------------------------------|                                                                                                   
RSGG2 | Random sister genus group 2 | Grafted as sister to a randomly selected genus from a pre-defined family            |                                                                                                   
--------------------------------------------------------------------------------------------------------------------------|
}
\examples{
# data(BirdTree_trees)
# data(BirdTree_tax)
# data(AvotrexPhylo)
# trees <- AvoPhylo(ctrees = BirdTree_trees, 
# avotrex = AvotrexPhylo, PER = 0.2, tax = BirdTree_tax, 
# Ntree = 1, n.cores = 1, cluster.ips = NULL)
}
\references{
Matthews et al. (IN REVIEW) The global loss of avian functional and phylogenetic diversity from extinctions in 
the Holocene and Late Pleistocene; Sayol et al. (IN PREP) The global loss of avian functional and phylogenetic diversity 
from extinctions in the Holocene and Late Pleistocene
}
