% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_AvotrexPhylo.R
\docType{data}
\name{AvotrexPhylo}
\alias{AvotrexPhylo}
\title{Instructions to build phylogenetic trees incorporating extinct bird species from AvoTrex dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 600 rows and 23 columns.
}
\usage{
data(AvotrexPhylo)
}
\description{
Instructions to build phylogenetic trees incorporating extinct bird species from AvoTrex dataset
}
\examples{
data(AvotrexPhylo)
}
\references{
To be updated
}
\keyword{datasets}
