% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_BirdTree_trees.R
\docType{data}
\name{BirdTree_trees}
\alias{BirdTree_trees}
\title{Example trees from Jetz et al. (2012)}
\format{
An object of class \code{multiPhylo} of length 5.
}
\usage{
data(BirdTree_trees)
}
\description{
Example trees from Jetz et al. (2012)
}
\examples{
data(BirdTree_trees)
}
\references{
Jetz, W., Thomas, G. H., Joy, J. B., Hartmann, K., & Mooers, A. O. (2012). The global diversity of birds in space and time. Nature, 491(7424), 444-448.
}
\keyword{datasets}
