% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesEx.R
\docType{data}
\name{treesEx}
\alias{treesEx}
\title{A set of three AvoPhylo trees}
\format{
An object of class 'multiAvophylo'
}
\usage{
data(treesEx)
}
\description{
A "multiAvophylo" object containing two AvoPhylo trees,
generated using the AvoPhylo function. Includes 610 extinct
species.
}
\examples{
data(treesEx)
}
\keyword{datasets}
